+++ src/05copyright.js
@@ -1,8 +1,8 @@
 //
 // alasql.js
 // AlaSQL - JavaScript SQL database
-// Date: 19.04.2015
-// Version: 0.0.48
+// Date: 21.04.2015
+// Version: 0.0.49
 // (ñ) 2014-2015, Andrey Gershun
 //