+++ src/50expression.js
@@ -112,8 +112,8 @@ yy.Op.prototype.toString = function() {
 	if(this.allsome) {
 		return this.left.toString()+" "+P(this.op)+" "+this.allsome+' ('+this.right.toString()+')';
 	}
-	if(this.op == '->') {
-		var s = this.left.toString()+"->";
+	if(this.op == '->' || this.op == '#') {
+		var s = this.left.toString()+this.op;
 //		console.log(this.right);
 		if(typeof this.right != 'string' && typeof this.right != 'number' ) s += '(';
 		s += this.right.toString();
@@ -172,6 +172,13 @@ yy.Op.prototype.toJavaScript = function(context,tableid,defcols) {
 		} else {
 			return this.left.toJavaScript(context,tableid, defcols)+'['+this.right.toJavaScript(context,tableid, defcols)+']';
 		}
+	};
+
+	if(this.op == '#') {
+		if(typeof this.right == "string") {
+			return 'alasql.databases[alasql.useid].objects['+this.left.toJavaScript(context,tableid, defcols)+']["'+this.right+'"]';
+		}		
+		// TODO - add other cases
 	}
 
 	if(this.op == 'IS') {