+++ test/test222.js
@@ -5,6 +5,10 @@ if(typeof exports === 'object') {
 	__dirname = '.';
 };
 
+if(false) {
+
+// Functionality discontinued
+// AG 20.04.2015
 describe('Test 222 TD TH syntax', function() {
     it("1. TD", function(done) {
         alasql('SELECT _ TD {className:"red"} FROM RANGE(1,2)');
@@ -13,3 +17,4 @@ describe('Test 222 TD TH syntax', function() {
 
 });
 
+}